% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mapped_courses.R
\name{get_mapped_courses}
\alias{get_mapped_courses}
\title{Get Mapped Courses}
\usage{
get_mapped_courses()
}
\value{
A data.frame with columns api_id, course_name_raw, city, state, trace_date, logo_url, course_name, and course_name_concat.
}
\description{
Returns a data frame of all mapped golf courses
}
