% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_course.R
\name{plot_course}
\alias{plot_course}
\title{Plot a golf course map by API ID}
\usage{
plot_course(api_id)
}
\arguments{
\item{api_id}{The course API ID (numeric or character)}
}
\value{
A ggplot object showing the course layout
}
\description{
Plot a golf course map by API ID
}
