% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hole.R
\name{plot_hole}
\alias{plot_hole}
\title{Plot a Single Golf Course Hole}
\usage{
plot_hole(api_id, hole_number)
}
\arguments{
\item{api_id}{Numeric or character. The API ID of the golf course.}

\item{hole_number}{Integer from 1 to 18. The hole number to plot.}
}
\value{
A ggplot object displaying the layout of the selected hole.
}
\description{
Generates a ggplot visualization of a specific hole of a golf course
using polygon data, based on the course's API ID and hole number.
}
