% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polygon_data.R
\name{get_polygon_data}
\alias{get_polygon_data}
\title{Get polygon data for a course}
\usage{
get_polygon_data(api_id)
}
\arguments{
\item{api_id}{The ID of the course}
}
\value{
An sf object with course polygons
}
\description{
Get polygon data for a course
}
